Refactor store domain types
