+++ src/components/TownSquare.tsx
@@ -3,7 +3,8 @@ import * as React from 'react';
 import tokenImg from './../assets/images/town/token.png';
 import { CharacterTokenParent } from './CharacterTokenParent';
 import { $$ROLES, CharacterTypes, Roles } from '../data/types';
-import { ISeatedPlayer, selectScript } from '../store/game/game-slice';
+import { selectScript } from '../store/game/game-slice';
+import { ISeatedPlayer } from '../store/types/player-types';
 import { useAppSelector } from '@/store/hooks';
 import { buildNightOrderIndex } from '../utils/nightOrder';
 import { Button } from '@/components/ui/button';
