+++ src/store/types/game-types.ts
@@ -1,0 +1,1 @@
+export type GameStates = 'idle' | 'reveal' | 'in-progress' | 'setup';
