+++ src/store/types/grimoire-types.ts
@@ -1,0 +1,16 @@
+import { Roles } from '../../data/types';
+import { ISeat } from './player-types';
+
+export interface IReminderTokens {
+    key: string;
+    source: number;
+    target: number;
+    isChanneled: boolean;
+}
+
+export interface IGrimoireSlice {
+    seats: ISeat[];
+    demonBluffs: [Roles, Roles, Roles] | undefined;
+    outOfPlay: Roles[];
+    reminderTokens: Record<string, IReminderTokens>;
+}
